several upgradades docker and scrapers
