+++ .gitignore
@@ -45,4 +45,3 @@ docker/secrets
 
 # 
 data/
-datavolume/
