+++ Dockerfile
@@ -5,11 +5,12 @@ RUN pip install --upgrade pip \
  && pip install selenium
 
 COPY main.py /opt/app/main.py
-COPY loop.py /opt/app/loop.py
 
 COPY src /opt/app/src
 COPY requirements.txt /opt/app/requirements.txt
 
 WORKDIR /opt/app
 RUN pip install -r requirements.txt
+
+
