--- loop.py
@@ -1,2 +1,0 @@
-while(True):
-    continue
