+++ main.py
@@ -29,13 +29,12 @@ parser.add_argument(
 )
 
 
-time.sleep(30)
 args = parser.parse_args()
 
 print(args.runtime)
 
 if 'flexicar' in args.site:
-    from src.es.flexicar import get_data_adds,get_links_cars,get_master
+    from src.scrapers.es.flexicar import get_data_adds,get_links_cars,get_master
     print("flexicar")
     print("master")
     get_master.main(args.runtime)
@@ -45,11 +44,11 @@ if 'flexicar' in args.site:
     get_data_adds.main(args.runtime)
 
 if 'mobile_de' in args.site:
-    from src.de.mobile_de import get_data_adds,get_links_cars,get_master
+    from src.scrapers.de.mobile_de import get_data_adds,get_links_cars,get_master
     print("mobile_de")
     print("master")
-    # get_master.main(args.runtime)
-    # print("links")
-    # get_links_cars.main(args.runtime)
-    # print("data")
+    get_master.main(args.runtime)
+    print("links")
+    get_links_cars.main(args.runtime)
+    print("data")
     get_data_adds.main(args.runtime)
