+++ src/scrapers/de/mobile_de/get_master.py
