+++ src/scrapers/utils.py
@@ -1,4 +1,5 @@
 import yaml
+from google.cloud import storage
 import pandas as pd
 import notion_df
 import os.path
